requirements.txt
numpy==1.16.6
Werkzeug==0.16.0
Flask==1.1.1
opencv_python==4.2.0.34
Pillow==7.2.0
tensorflow>=2.2.0
gunicorn>=19.9.0
